- laser timings now adaptable with some variables

+++ Software/Firmware/OpenExposer/LaserDriver.h
@@ -8,7 +8,7 @@
 extern uint8_t exposing_done;
 
 typedef union {
-    uint32_t positions[LASER_POSITIONS_BUFFER_SIZE];
+    long positions[LASER_POSITIONS_BUFFER_SIZE];
     uint16_t laser_timings[LASER_POSITIONS_BUFFER_SIZE];
 } laser_buffer_type;
 
@@ -17,7 +17,7 @@ void laser_on();
 void laser_off();
 void expose_line(int time);
 void set_exposing_cycles(uint8_t cycles);
-void fill_laser_buffer(long distance);
+void fill_laser_buffer(int32_t distance);
 void create_test_pattern();
 void convert_positions_to_timings();
